lab/gcp/19-12-2025-meeting/func_stat/requirements.txt
google-cloud-firestore==2.22.0

# Se contiene una HTTP function 
flask==2.3.3